Run the tests with existing image
When ran in functest on arm, the doctor suite fails because the x86
image is hardocded in the source files. This change adds the option to
run the doctor suite with an existing image as opposed to downloading
and creating it.

The image name is taken from an enviroment variables (if it is set).
When ran with functest, the env variable is set in doctor.py file.

+++ tests/run.sh
@@ -11,7 +11,8 @@
 [[ "${CI_DEBUG:-true}" == [Tt]rue ]] && set -x
 
 IMAGE_URL=https://launchpad.net/cirros/trunk/0.3.0/+download/cirros-0.3.0-x86_64-disk.img
-IMAGE_NAME=cirros
+#if an existing image name is provided in the enviroment, use that one
+IMAGE_NAME=${IMAGE_NAME=:-cirros}
 IMAGE_FILE="${IMAGE_NAME}.img"
 IMAGE_FORMAT=qcow2
 VM_NAME=doctor_vm1
@@ -45,6 +46,9 @@ if [[ ! "$SUPPORTED_INSPECTOR_TYPES" =~ "$INSPECTOR_TYPE" ]] ; then
     echo "ERROR: INSPECTOR_TYPE=$INSPECTOR_TYPE is not supported."
     exit 1
 fi
+
+use_existing_image=false
+openstack image list | grep -q " $IMAGE_NAME " && use_existing_image=true
 
 get_installer_ip() {
     if [[ "$INSTALLER_TYPE" == "apex" ]] ; then
@@ -252,8 +256,11 @@ get_consumer_ip() {
 }
 
 download_image() {
-    [ -e "$IMAGE_FILE" ] && return 0
-    wget "$IMAGE_URL" -o "$IMAGE_FILE"
+    #if a different name was provided for the image in the enviroment there's no need to download the image
+    if [[ "$use_existing_image" == false ]] ; then
+        [ -e "$IMAGE_FILE" ] && return 0
+        wget "$IMAGE_URL" -o "$IMAGE_FILE"
+    fi
 }
 
 register_image() {
@@ -531,7 +538,10 @@ cleanup() {
 
     image_id=$(openstack image list | grep " $IMAGE_NAME " | awk '{print $2}')
     sleep 1
-    [ -n "$image_id" ] && openstack image delete "$image_id"
+    #if an existing image was used, there's no need to remove it here
+    if [[ "$use_existing_image" == false ]] ; then
+        [ -n "$image_id" ] && openstack image delete "$image_id"
+    fi
     openstack role remove "$DOCTOR_ROLE" --user "$DOCTOR_USER" \
                               --project "$DOCTOR_PROJECT"
     openstack project delete "$DOCTOR_PROJECT"
